requirement.txt
pymongo==3.9.0
pandas==0.25.3
tabulate==0.8.6
psycopg2==2.8.4
plotly==4.4.1
selenium==3.141.0
xlrd==1.1.0
pyyaml==5.2